feat: reservas backend funcional con JWT, integración Flutter pendiente

+++ backend/src/main/java/com/farfala/backend/Reserva/Reserva.java
@@ -14,9 +14,9 @@ public class Reserva {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
-    private Integer usuarioId;
+    private Long usuarioId;
 
     @ManyToOne
     @JoinColumn(name = "clase_id", nullable = false)
@@ -27,7 +27,7 @@ public class Reserva {
     private String hora;
 
     // ✅ Constructor personalizado SIN el id, para crear la reserva desde el controlador
-    public Reserva(Integer usuarioId, Clase clase, String fecha, String hora) {
+    public Reserva(Long usuarioId, Clase clase, String fecha, String hora) {
         this.usuarioId = usuarioId;
         this.clase = clase;
         this.fecha = fecha;
